units (with model mapper error)
error on parsing actual comission

+++ aviasales-common/src/main/java/ru/aviasales/common/dto/request/SalesUnitReq.java
@@ -26,6 +26,6 @@ public class SalesUnitReq {
     private Long categoryId;
 
     @Schema(description = "Процент комиссии")
-    private Double commissionPercent;
+    private Long commissionPercent;
 
 }
